Work in progress on updating the app for iOS 8 & 9

+++ Windandtides/Controllers/WindandtidesViewController.h
@@ -22,6 +22,6 @@
 
 - (IBAction)reloadWebViews:(UIButton *)button;
 
-- (void)loadWebView:(int)tabIndex;
+- (void)loadWebView:(long)tabIndex;
 
 @end
